Resolved minor errors and code cleanup
