+++ lib/Screens/display_screen.dart
@@ -1,35 +1,13 @@
-import 'package:evil_icons_flutter/evil_icons_flutter.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_slidable/flutter_slidable.dart';
-import 'package:nih_laundro/main.dart';
+import 'package:nih_laundro/Widgets/reset_button.dart';
+import 'package:nih_laundro/Widgets/save_button.dart';
 import 'package:nih_laundro/model/model.dart';
 import 'package:redux/redux.dart';
-import 'package:nih_laundro/redux/actions.dart';
 
 class DisplayScreen extends StatelessWidget {
   final Store<AppState> store;
 
   const DisplayScreen({Key key, this.store}) : super(key: key);
-
-  int getTotal() {
-    int total = 0;
-
-    for (int i = 0; i < store.state.clothes.length; i++) {
-      total += store.state.clothes[i].count * store.state.clothes[i].price;
-    }
-
-    return total;
-  }
-
-  void dispatchOnSaveClothes() {
-    store.dispatch(SetPrefsVariable());
-    store.dispatch(SaveClothesAction());
-  }
-
-  void dispatchResetCount() {
-    store.dispatch(ResetPrefsVariable());
-    store.dispatch(ResetCount());
-  }
 
   Widget createWidget(int i) {
     if (store.state.clothes[i].count != 0)
@@ -102,71 +80,14 @@ class DisplayScreen extends StatelessWidget {
                     ],
                   ),
                 ),
-                Slidable(
-                  actionPane: SlidableDrawerActionPane(),
-                  actionExtentRatio: 0.25,
-                  child: Container(
-                    height: 45.0,
-                    width: MediaQuery.of(context).size.width * 1,
-                    color: Color(0xFF2282B9),
-                    child: Row(
-                      children: <Widget>[
-                        Icon(EvilIcons.chevron_right,
-                            size: 25.0, color: Colors.white),
-                        Icon(EvilIcons.chevron_right,
-                            size: 25.0, color: Colors.white),
-                        Icon(EvilIcons.chevron_right,
-                            size: 25.0, color: Colors.white),
-                        Center(
-                          child: Text(
-                            'Total : Rs. ${getTotal().toString()} + Miscellanous',
-                            style: TextStyle(
-                              color: Colors.white,
-                              fontWeight: FontWeight.bold,
-                              fontSize: 20.0,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  actions: <Widget>[
-                    IconSlideAction(
-                      caption: 'Save',
-                      color: Color(0xFF4EB44B),
-                      icon: Icons.check,
-                      foregroundColor: Colors.white,
-                      onTap: () => dispatchOnSaveClothes(),
-                    ),
-                  ],
+                SaveButton(
+                  store: store,
                 ),
                 SizedBox(
                   height: 5.0,
                 ),
-                Container(
-                  width: MediaQuery.of(context).size.width * 1,
-                  height: 45.0,
-                  child: FlatButton(
-                    color: Color(0xFF2282B9),
-                    child: Text(
-                      'Reset',
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontWeight: FontWeight.bold,
-                        fontSize: 20.0,
-                      ),
-                    ),
-                    onPressed: () {
-                      dispatchResetCount();
-                      Navigator.of(context).pushReplacement(
-                        new MaterialPageRoute(
-                          builder: (BuildContext context) => HomePage(
-                            store: store,
-                          ),
-                        ),
-                      );
-                    },
-                  ),
+                ResetButton(
+                  store: store,
                 ),
               ],
             ),
